modules/cameraCapture/requirements.txt
azure-iot-device~=2.7.0
requests
flask
opencv-python
imutils